src/requirements.txt
numpy==1.18.2
Pillow==7.1.2
pytest==5.4.3
scikit-learn==0.23.1
scipy==1.4.1
sklearn==0.0
